cloud/storage: fix CopyObject - note breaking API change
If object attributes are sent to the Objects.Copy call but don't
include ContentType, the GCS API returns an error.
Passing 'nil' causes the GCS API to copy the metadata from the
source file to the destination file, including the ContentType.

Note that this changes the API for CopyObject.
Calls must now provide a "destName" argument and the attrs argument
can now be nil, in which case the source object's properties will
be copied to the destination object.

Change-Id: Ib97b02e9f12cc43b45584fababbcb95e5c960651