storage: enforce valid UTF8 object names.
In Writer, fail to write objects with invalid UTF8 in their names.
In ObjectHandle, return an error for Attrs, Delete, NewReader, and
Update if the object has invalid UTF8 in its name.
In CopyObject, check both srcName and destName for valid UTF8.

Increase the documentation for valid bucket and object names.

This fixes #111.

Change-Id: Ibefa7ab308667f96283ba00f58bc21436a6e9342
Reviewed-on: https://code-review.googlesource.com/3892
Reviewed-by: Michael McGreevy <mcgreevy@golang.org>