bundler: new package supporting bundling (batching)
Sometimes it is helpful to collect multiple user items (e.g. log
entries, pubsub messages) and periodically handle them as a group. The
Bundler in this package supports bundling by both time (e.g. every
second) and number of items.

This package will be of immediate use to the logging client. The pubsub
client should also be able to use it.

Change-Id: Ib77d06d0818f7cfcbf3ebf2ea66fb239f8947039
Reviewed-on: https://code-review.googlesource.com/5090
Reviewed-by: Michael McGreevy <mcgreevy@golang.org>