bigquery: add RowIterator, to replace Iterator
This CL introduces RowIterator, which follows the standard iterator
pattern. It also contains the replacements for Job.Read and Table.Read
that will return it, but makes no breaking changes.

The main surface change is that ReadOptions will go away:

- RecordsPerRequest becomes PageInfo().MaxSize.

- There never was a PageToken read option. Now PageInfo().Token can be
  set as usual.

- It no longer makes sense to have a single StartIndex option, when the
  other two are fields. So StartIndex is a field on RowIterator.

The implementation is a bit clumsy, out of a desire to leave the
existing XXXConf types and service methods mostly unchanged. After the
old iterator is removed, we should be able to simplify.

Change-Id: Ief1fa93dd93ce4cd8006c8c9cd14ec5f3559e8b0
Reviewed-on: https://code-review.googlesource.com/7612
Reviewed-by: Michael McGreevy <mcgreevy@golang.org>