pubsub: implement on top of generated client
The pubsub client uses the generated client, and therefore gRPC.

This is not a breaking change, but it may have performance implications
if you create many pubsub.Clients. The current version of gRPC does not
reuse connections, so each Client manages its own TCP connection. You
should create clients sparingly and call the Close method on them when
done (although closing on program termination is not necessary).

The biggest change was in the iterators. I made the smallest changes
that kept everything working. There will be more changes soon as
the generated code adopts the latest iterator design, and that
trickles into this client.

Change-Id: I6c4c1f4b50b52ff8f1d1ff403e2979cd81c6c7ab
Reviewed-on: https://code-review.googlesource.com/7911
Reviewed-by: Ross Light <light@google.com>