storage: support field deletion in ObjectHandle.Update
Add ObjectAttrsToUpdate, which has all and only the writeable fields of an
object.

Introduce a new package "optional" with types String, etc., all defined
to be interface{}.  Use these to define the fields of primitive type in
ObjectAttrsToUpdate, so we can distinguish an explicit zero value from
the field being unset.

For types that can be nil, like map[string]string, use the type itself
and require a zero value to delete.

For setting values (the only case we currently support), usage looks
identical:

    obj.Update(ctx, ObjectAttrsToUpdate{ContentType: "text/html"})

(Though it is not type-checked at compile-time.)

To clear a value, explicitly set it to the zero value:

    obj.Update(ctx, ObjectAttrsToUpdate{
	ContentType: "",		// set ContentType to "" (same
					// as deleting it)
        Metadata: map[string]string{},	// delete metadata
    })

Change-Id: Ic9e769914cb45aebb75e5769bec41fb22ca5a5f8
Reviewed-on: https://code-review.googlesource.com/7970
Reviewed-by: Dave Day <djd@golang.org>