bigquery: pass Schema to ValueLoader
BREAKING CHANGE.

The standard iterator design makes it awkward to use the returned schema
with a ValueLoader. Previously, one could call it.Next, then call
it.Schema, then use the schema to affect the ValueLoader before calling
it.Get. The new design, by combining Next and Get, makes this awkward.

Instead, we pass the schema to ValueLoader.Load.

As a further benefit, this simplifies RowIterator, because there is no
longer the need for a Schema method.

Fixes #324.

Change-Id: Iceb470ac9e8ccb2039d992284161ac8c21660ce0
Reviewed-on: https://code-review.googlesource.com/8553
Reviewed-by: Ross Light <light@google.com>