bigquery: ValueLoader for structs
Implement structLoader, a ValueLoader for structs. A structLoader is
used whenever a pointer to a struct is passed to RowIterator.Next.

A structLoader uses reflection to populate a struct's fields with the
columns of a row.

It precomputes as much as possible about the schema for performance,
"compiling" the (struct type, schema) pair into an efficient sequence of
operations. The resulting performance gain over using reflection for
each row is signficant.

Wherever possible, struct loading tries to behave like encoding/json
unmarshaling.

Fixes #399.

Change-Id: I8a674171a5cb24e91046b0ee818da7bafb350f23
Reviewed-on: https://code-review.googlesource.com/9175
Reviewed-by: Michael Darakananda <pongad@google.com>
Reviewed-by: Ross Light <light@google.com>