bigquery: more types for RowIterator.Next
Breaking change, but only for programs that mention the signature of
RowIterator.Next. Call sites will not break.

Change the argument type of RowIterator.Next to interface{}, to simplify
its use. Continue to accept a ValueLoader, but also accept a *[]Value
or *map[string]Value.

With this change, there is no reason to export ValueMap, so I unexported it.

This change becomes much more valuable when loading from structs is
introduced (CL forthcoming). For efficiency, a struct loader must exist
across calls to Next, which differs from the pattern usually used with
ValueLoaders. It's more convenient for the user if that machinery is
hidden in the iterator.

Part of #399.

Change-Id: I0366be1a6af020a49945b7ed9c08fee0f29cbb0f
Reviewed-on: https://code-review.googlesource.com/9173
Reviewed-by: Jun Mukai <mukai@google.com>