fields: add a cache
Add a Cache type, which caches the results of calling Fields on a type.

This can save a lot of time if the type's fields are retrieved many times,
such as when serializing many values of the type.

The optimized locking is taken directly from
https://go.googlesource.com/go/+/go1.7.3/src/encoding/json/encode.go,
where it was added to improve performance.

Change-Id: Ifcd1ab2b115c9ae693ef64a42018c3688d90f77f
Reviewed-on: https://code-review.googlesource.com/9510
Reviewed-by: Jun Mukai <mukai@google.com>