all: refactor authn/authz examples, add notices to subpackages
Refactored auth examples in the root package and modified
godoc for subpackages to add an Authorization section referring
to the root package. I am not sure this is the entire list of
subpackages that needs to be changed (some of the package godocs
were in auto-generated files).

Opening mostly to discuss if we can add pointer for auth{n,z}
in the subpackages.

Change-Id: I39b5c21b7e74e96eadb5dd1d8165fcae1d8917b1
Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>
Reviewed-on: https://code-review.googlesource.com/11230
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Sarah Adams <shadams@google.com>