cmd/go-cloud-debug-agent: include labels when registering debuggee.
The Debugger UI uses the module and version labels to show as the name
of the current running debuggee, and uses the minorversion to tell if to
combine multiple deployments of the same version of debuggee together.
This change will solve the current UI bug where use may see two
"version" on the debug console.

Change-Id: I9952b094548715002e67b7a056be5cce8cc4eac7
Reviewed-on: https://code-review.googlesource.com/11232
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Herbie Ong <herbie@google.com>