pubsub: fix flaky streaming retry test
When the stream of a streaming pull is closed by the server and there
is a client message in flight, the client message is lost. This caused
our streaming pull retry test to be flaky.

Fixed by adding a small delay to the server before returning an error.

Note that this is totally an artifact of our tests: at the end of the
interaction we compare what the server received with what it ideally
should have received.  With the actual service, dropped acks are a
real possibility.

Change-Id: I2e0af67d52371779d56ac6cd65b93d133bac3227
Reviewed-on: https://code-review.googlesource.com/11376
Reviewed-by: Michael Darakananda <pongad@google.com>