container/container.go: InitialNodeCount is always zero
The InitialNodeCount in the cluster(s) response is always zero so it is
not a useful metric to return. The more interesting metric is the
CurrentNodeCount. A similar argument could be made for the
InitialClusterVersion but this should be addressed in a separate commit
if it is decided that it should be changed

Change-Id: Ic2b8ebafb59324da75e7ae3192c2fbb6fc121896
Reviewed-on: https://code-review.googlesource.com/12450
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>