datastore: replay on -short
Instead of skipping integration tests on -short,
replay from a file if the file exists.

The command

    go test cloud.google.com/go/datastore -record

writes the file "datastore.replay" to the datastore directory
(regardless of where it is run).

Add this file to a CL to update the recording.

Subsequent "go test -short" commands will use the file to replay
integration tests.

Change-Id: I02c1ce15f9ddf4ce71d827006ca1f952e77ea9ce
Reviewed-on: https://code-review.googlesource.com/14613
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Ross Light <light@google.com>