bigquery: prefer query error to missing dest table error
We assumed a returned Job would have a destination table, and checked
for its absence. That is almost right: it may not have a destination
table if the query has an error. So move the nil check later.

One wrinkle: we need to get the project ID from the client if there is
no destination table.

Fixes #642.

Change-Id: Ibdb20ba68393f13b63ffff6c0a63c0d6e0c24dad
Reviewed-on: https://code-review.googlesource.com/13610
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Sai Cheemalapati <saicheems@google.com>