rpcreplay: file format and I/O
Recording writes a file that replaying will read. This CL describes
the file format and implements reading and writing. I use a proto
for Entry so the files can be accessed from other languages.

The messageOrError type may seem superfluous, but it will turn
out to be useful to support streaming.

Change-Id: I9d2a8892d7c213aba8f0692c56ccce717d6a0d92
Reviewed-on: https://code-review.googlesource.com/13750
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Sarah Adams <shadams@google.com>