many: Remove reflect.DeepEqual, part 1
Replace reflect.DeepEqual in a few smaller packages. Larger packages
to be tackled in later CLs.

We will use Joe Tsai's excellent new cmp package. (Yay Joe!)

Make a repo-wide wrapper for cmp.Equal in internal/testutil that
handles protos and NaNs, since we'll always want to compare them
correctly. It's also nice that the package name "testutil" makes it
clear that this is for testing only.

Change-Id: I05f22da6c44b0ce1d2113778af1cabaa2136a3fb
Reviewed-on: https://code-review.googlesource.com/14614
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>
Reviewed-by: Ross Light <light@google.com>