profiler: symbolize profiles
runtime/pprof in some go versions, for example go1.8.x, doesn't symbolize
profiles. We need to symbolize the profiles in this case.

Most of the code are taken from https://go-review.googlesource.com/c/34192/. We
add logic to keep the root location of a sample if all locations are
trimmed.

Change-Id: I3e98e3cecab7866f8031dd533c433f0407dca59e
Reviewed-on: https://code-review.googlesource.com/14850
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Alexey Alexandrov <aalexand@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>