bigquery: support update of Dataset.DefaultTableExpiration
Allow users to change or delete the DefaultTableExpiration property
of a Dataset.

We have to distinguish "do not update" from "delete". To do so,
we make an optional type for Duration, interpret the default value (nil)
as "do not update", and treat a zero duration (otherwise illegal) as "delete".
Note that the bq command-line tool makes the same choice: you write

  bq update --default_table_expiration 0

to delete. We also do this in storage.ObjectHandle.Update, where we
have a Metadata field that is a map. If the map is nil, nothing
happens; if you set it to an empty map, the metadata is deleted.

Change-Id: I3ffdf580cd84c12ea2527d49ddd61dc0b04a53c7
Reviewed-on: https://code-review.googlesource.com/15570
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>