logging: Update logging.ErrOverflow to point to bundler.ErrOverflow
The logging package defines an ErrOverflow sentinel error but never uses
it. The documentation of the OnError client field and the
BufferedByteLimit option both inaccurately indicate that it is returned
by the logging package.

The error was previously used and stopped being used when the bundling
logic was broken out into a separate bundler package. The bundler
package defines its own ErrOverflow sentinel error.

Update "logging.ErrOverflow" to point to "bundler.ErrOverflow". This
makes it so the documentation is accurate and prevents any existing code
relying on the previous behaviour from breaking.

Do the same with the "bundler.ErrOversizedItem" error for the sake of
completeness.

Change-Id: Ic004610ec2e14ccdec45eb30966596c27ed4deb7
Reviewed-on: https://code-review.googlesource.com/16130
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>