bigquery: support jobs.list
Add Client.Jobs, which returns an iterator over JobInfos.

Support filtering by project ID, all users, and job state.

A JobInfo includes a Job, holding the immutable parts of a job (ID,
config), and a JobStatus. A later CL will expose the job config.

At present we always read the job config. The "minimal" view, which
omits the config, is a performance optimization. We can include it
when we get feedback that the full view is too slow.

Change-Id: Id90ee0efdbab6c5123fdd0110ab6c3ad80d02122
Reviewed-on: https://code-review.googlesource.com/16910
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>
Reviewed-by: Jeff Jolma <jjolma@google.com>
Reviewed-by: Tim Swast <swast@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>