storage: add notifications
Support PubSub notification, which sends PubSub messages when
certain events occur on a bucket.

The new methods support adding and removing notifications, and
listing all of a bucket's current notifications.

One unusual feature of the design: BucketHandle.Notifications returns
a map from notification ID to Notification, instead of a slice or an
iterator. We don't use an iterator because the underlying RPC,
storage.notifications.list, returns a complete list of notifications
instead of using pagination, indicating that the number of
notifications for a bucket is expected to be small. We return a map
instead of a slice so callers can easily get a particular notification
by ID. This obviates the need for a separate "get" method (and indeed
we do not implement storage.notifications.get).

Change-Id: If3f9cc4fdce5f64f9ea9a10d67ba12376b8dda36
Reviewed-on: https://code-review.googlesource.com/18450
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Ross Light <light@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>