firestore: support set/merge with Delete
The data given DocRef.Set can contain Delete sentinels if a merge option
is provided. The behavior is like Update, except that there is no precondition.

To exploit the similarity between set/merge and update, in both cases
we compute a set of fieldpath-value pairs and use them to generate the
request.

This CL also includes additional conformance tests for the new behavior.

Change-Id: I0546753e1be8e11ff2c1edab3f316d134393fff6
Reviewed-on: https://code-review.googlesource.com/21950
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>
Reviewed-by: Sebastian Schmidt <mrschmidt@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>