firestore: remove UpdateMap and UpdateStruct
BREAKING CHANGES:
- UpdateMap removed
- UpdateStruct removed
- UpdatePaths renamed to Update
- FieldPathUpdate renamed to Update
  - Path field changed type to string
  - New FieldPath field

UpdateMap is similar to Set with merge, but different in confusing ways. If

  m := map[string]interface{}{"a.b.c": 1}

then

  Set(m, MergeAll)

would update the top-level key "a.b.c", but

  UpdateMap(m)

would split "a.b.c" on dots and replace the nested value.

Removing UpdateMap reduces the potential for confusion.

Without UpdateMap, it doesn't make sense to keep UpdateStruct around. It is
extremely close to Set with a list of merge paths.

We keep the former UpdatePaths so users can still perform an Update: changing selected
paths with the precondition that the document exists. But without the other
two methods, we can rename it to Update.

We also rename FieldPathUpdate to Update just to make it easier to write.

These changes remove a lot of "convenience," but it remains to be seen whether
users will really miss anything. We can always add back convenience, but once
we exit beta we can never remove it.

Change-Id: Iabe086c61085b9c55f8fcce89c7ba4a38bc3d52e
Reviewed-on: https://code-review.googlesource.com/20150
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>