internal/snipdoc: better maintenance of doc.go files
Snipdoc is a simple tool that prevents code rot in doc.go files.

snipdoc.awk merges a text template with Go code. The Go code can
be compiled to verify that it still works.

I've been using snipdoc in the firestore client. This CL just pulls
it into a common place and documents it.

Change-Id: I2f5a0936f614fab28894ca0313f53dff392861ae
Reviewed-on: https://code-review.googlesource.com/20170
Reviewed-by: Michael Darakananda <pongad@google.com>