bigtable/bttest: Fix a race condition with MutateRows in the emulator
The previous locking scheme left a small window open where new rows
had been appended to the row list but not sorted. During this time
the new rows would be incorrectly included in reads.

In the future we need to move away from this representation of
data in the emulator. Rows should really be stored in a sorted data
structure with O(log(n)) insertions.

Change-Id: Ifaa45adc8253e1ba4e73d4e1c3c900a064077112
Reviewed-on: https://code-review.googlesource.com/21553
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>