pubsub: resumable stream implementation
Provide a type that behaves like a StreamingPullClient, but re-opens the stream
on a retryable error.

This borrows a key idea from https://code-review.googlesource.com/c/gocloud/+/21850:
we can avoid complex locking if we compare instances of the stream to verify
we're using the most up-to-date one.

Also, tweak the timeout test to avoid flakiness.

Change-Id: I053834c024d54e24edcd1356f038203c4c494bde
Reviewed-on: https://code-review.googlesource.com/21930
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>