firestore: add DocumentIterator.Stop
We need to clean up gRPC streams if we don't read them to the end. The
way to do that is to cancel the context that we give to the streaming
RPC.

To make sure the cancel function gets called, we add a Stop method to
the iterator.

We call Stop ourselves in DocumentIterator.GetAll, to enable chaining:

   q.Documents(ctx).GetAll()

Although this CL is not technically a breaking change, it does demand
a change of existing users: they need to call the Stop method.

Change-Id: I76dabb72ac6fdb9af0e5da34a102c10f5cd6ab99
Reviewed-on: https://code-review.googlesource.com/25751
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>